requirements.txt
rasa==3.6.21
rasa-sdk==3.6.2
requests
python-dotenv
SQLAlchemy<2.0
